factored out user interaction and log output
For log output, the logging module is used.

For user interaction, a new class TextUserInteraction is added that
implements all ways main.py needs to interact with the user.

Any object quacking the same can be substituted for this, thereby enabling
the user or a program to interact with rapid in a different way.

Tiny behaviour changes:
- slightly different formatting / wording of some messages
- selecting multiple options is now possible (one or all => one, many or all)