add -r/--regex option to use regex instead of substring matches
Should work for all commands that take a searchterm.
(*pin, *install, list*, etc.)

The regexes are case insensitive and not anchored by default.
(Use ^ and $ to anchor to begin and end of string respectively.)

Example: rapid -r list-tags '^(?!ba).*:(latest|stable|test)$'
This displays all tags ending with ':latest', ':stable' or ':test',
which do not start with the string 'ba'.

Closes gh-2